Initial Commit
Merged aita.py and subreddit_drama.py into one with slash command grouping set. To be tested.
