+++ extension/reddit_commands.py
@@ -1,14 +1,13 @@
-"""
-Author: Kai (Not Lance (Not Russell))
-"""
-
 import config
 import discord
 from discord.ext import commands
 from discord.ui import Button, View
 from discord import ButtonStyle
 from discord.commands import slash_command
+from util.reddit_scraper import get_subreddit_drama
 from util.reddit_scraper import get_aita
+
+reddit_set = discord.SlashCommandGroup('reddit', 'Commands which retrieve top posts from subreddits', guild_ids=[config.GUILD_ID])
 
 class AitaView(View):
     def __init__(self, post):
@@ -56,8 +55,7 @@ class AITA(commands.Cog):
     def __init__(self, bot: commands.Bot):
         self.bot = bot
 
-    @slash_command(guild_ids=[config.GUILD_ID],
-        description="Is OP the asshole?")
+    @reddit_set.command(description='What\'s trending on r//AmITheAsshole')
     async def aita(self, ctx: discord.ApplicationContext):
         try:
             await ctx.defer()
@@ -79,5 +77,38 @@ class AITA(commands.Cog):
 
         await ctx.respond(embed=embed, view=AitaView(post))
 
+class DramaView(View):
+    def __init__(self, post):
+        super().__init__()
+        self.post = post
+
+class DRAMA(commands.Cog):
+    def __init__(self, bot: commands.Bot):
+        self.bot = bot
+
+    @reddit_set.command(description='What\'s trending on r//SubredditDrama')
+    async def subreddit_drama(self, ctx: discord.ApplicationContext):
+        try:
+            await ctx.defer()
+            post = await get_subreddit_drama()
+        except Exception as e:
+            return await ctx.respond('Unable to find a spicy story :/', ephemeral=True)  
+        
+        # must truncate embed descriptions to 4096 chars
+        if len(post['desc']) > 4096:
+            post['desc'] = post['desc'][:4096-3] + '...' 
+
+        embed = discord.Embed(
+            title=post['title'],
+            url=f"https://reddit.com{post['link']}",
+            description=post['desc']
+        )
+        embed.set_footer(text=f"{post['upvotes']} upvotes")
+        embed.set_author(name=post['author'])
+
+        await ctx.respond(embed=embed, view=DramaView(post))
+
+
 def setup(bot):
     bot.add_cog(AITA(bot))
+    bot.add_cog(DRAMA(bot))
